v4.1.4
Polls every 5 s, and when face leaves camera on front camera it stops toasting.
Unf, it has the following bugs:
1. once camera is switched and it was previously toasting, it will continue toasting
2. continues toasting once user leaves app
Now I only realize it is the same reason that causes both events to occur. That the app doesn't reset upon both events: when camera is changed.i.e. there's a bug that involves not being able to listen to facedetection events any more. How to settle them is by 
#1 setting protection = 0 upon switching camera (unf I tried this - maybe i need to try harder) & exit
#2 perhaps its prot cannot be set to 0 upon these two events & only solution is to turn off runnable handler. But how? By breaking - i.e. break statement.